# assets/docs/rider/v1/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "### 说明\n接口采用非标准Restful API，所有http返回代码均为`200`，当返回为非`200`时应为network错误，需要及时排查。\n\u003cbr\u003e\n接口返回说明查看 **[返回](#返回)**\n\u003cbr\u003e\n图片/附件前缀 `https://cdn.auroraride.com/`\n\n\u003cbr /\u003e\n\n\u003cbr /\u003e\n\n### 接口编号\n\n第一位代表接口端分类\n\n- M 管理端\n- R 骑手端\n- E 门店端\n- C 通用\n\n第二位代表子分类（36进制）\n\n后三位代表子编号（10进制）\n\n\u003cbr /\u003e\n\n\u003cbr /\u003e\n\n\n### 认证\n项目接口使用简单认证，认证方式为`header`中添加对应的认证`token`\n|  header   |  类型  |  接口  |\n| :-----: | :----: | :--: |\n|  X-Rider-Token   |  string   |  骑手API  |\n| X-Manager-Token | string |  后台API  |\n|  X-Employee-Token   | string |  员工API  |\n\n\u003cbr /\u003e\n\n\u003cbr /\u003e\n\n### 返回\n\n一个标准的返回应包含以下结构\n\n|  字段   |  类型  |  必填  |  说明  |\n| :-----: | :----: | :--: | :--: |\n|  code   |  int   |  是  |  返回代码  |\n| message | string |  是  |  返回消息  |\n|  data   | object |  是  |  返回数据  |\n\n`code`代码取值说明\n\n| 十进制 | 十六进制 | 说明 |\n| :----: | :------: | :--: |\n| 0  |  0x000  | 请求成功 |\n| 256 |  0x100  | 请求失败 |\n| 512 |  0x200  | *需要认证(跳转登录) |\n| 768 |  0x300  | *用户被封禁 |\n| 1024 |  0x400  | 资源未获 |\n| 1280 |  0x500  | 未知错误 |\n| 1536 |  0x600  | *需要实名 |\n| 1792 |  0x700  | *需要验证 (更换设备, 需要人脸验证) |\n| 2048 |  0x800  | *需要联系人 |\n| 2304 |  0x900  | 请求过期 |\n\n\u003e 当返回值是`1792(0x700)需要人脸验证`或`1536(0x600)需要实名`的时候`data`返回值为`{\"url\": \"string\"}`, 直接跳转url\n\n\n比如：\n\u003e 默认成功返回\n```json\n{\n  \"code\": 0,\n  \"message\": \"OK\",\n  \"data\": {\n    \"status\": true\n  }\n}\n```\n",
...
    },
    "basePath": "/",
    "paths": {
...
        "/rider/v1/setting/app": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Setting - 设置"
                ],
                "summary": "R6001 获取APP设置",
                "operationId": "RiderSettingApp",
                "parameters": [
                    {
                        "type": "string",
                        "description": "骑手校验token",
                        "name": "X-Rider-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/model.SettingRiderApp"
                        }
                    }
                }
            }
        },
        "/rider/v1/setting/question": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Setting - 设置"
                ],
                "summary": "R6002 获取常见问题",
                "operationId": "RiderSettingQuestion",
                "parameters": [
                    {
                        "type": "string",
                        "description": "骑手校验token",
                        "name": "X-Rider-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.SettingQuestion"
                            }
                        }
                    }
                }
            }
        },
...
        "/rider/v1/wallet/coupons": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet - 钱包"
                ],
                "summary": "R9004 优惠券列表",
                "operationId": "RiderWalletCoupons",
                "parameters": [
                    {
                        "type": "string",
                        "description": "骑手校验token",
                        "name": "X-Rider-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "enum": [
                            0,
                            1,
                            2
                        ],
                        "type": "integer",
                        "description": "查询类别 0:可使用 1:已使用 2:已过期",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.CouponRider"
                            }
                        }
                    }
                }
            }
        },
        "/rider/v1/wallet/overview": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet - 钱包"
                ],
                "summary": "R9001 钱包概览",
                "operationId": "RiderWalletOverview",
                "parameters": [
                    {
                        "type": "string",
                        "description": "骑手校验token",
                        "name": "X-Rider-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/model.WalletOverview"
                        }
                    }
                }
            }
        },
        "/rider/v1/wallet/pointlog": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet - 钱包"
                ],
                "summary": "R9002 积分日志",
                "operationId": "RiderWalletPointLog",
                "parameters": [
                    {
                        "type": "string",
                        "description": "骑手校验token",
                        "name": "X-Rider-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "当前页, 从1开始, 默认1",
                        "name": "current",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数据, 默认20",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.PaginationRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.PointLogListRes"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/rider/v1/wallet/points": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet - 钱包"
                ],
                "summary": "R9003 积分详情",
                "operationId": "RiderWalletPoints",
                "parameters": [
                    {
                        "type": "string",
                        "description": "骑手校验token",
                        "name": "X-Rider-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/model.PointRes"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "model.CouponRider": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "金额",
                    "type": "number"
                },
                "code": {
                    "description": "券码",
                    "type": "string"
                },
                "expiredAt": {
                    "description": "到期时间",
                    "type": "string"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "usedAt": {
                    "description": "使用时间",
                    "type": "string"
                }
            }
        },
...
        "model.Pagination": {
            "type": "object",
            "properties": {
                "current": {
                    "description": "当前页",
                    "type": "integer"
                },
                "pages": {
                    "description": "总页数",
                    "type": "integer"
                },
                "total": {
                    "description": "总条数",
                    "type": "integer"
                }
            }
        },
        "model.PaginationRes": {
            "type": "object",
            "properties": {
                "items": {
                    "description": "返回数据"
                },
                "pagination": {
                    "description": "分页属性",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Pagination"
                        }
                    ]
                }
            }
        },
...
        "model.PointLogListRes": {
            "type": "object",
            "properties": {
                "after": {
                    "description": "变动后",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "modifier": {
                    "description": "操作人",
                    "type": "string"
                },
                "plan": {
                    "description": "订单骑士卡",
                    "type": "string"
                },
                "points": {
                    "description": "积分",
                    "type": "integer"
                },
                "reason": {
                    "description": "原因",
                    "type": "string"
                },
                "time": {
                    "description": "时间",
                    "type": "string"
                },
                "type": {
                    "description": "类别",
                    "type": "string"
                }
            }
        },
        "model.PointRes": {
            "type": "object",
            "properties": {
                "locked": {
                    "description": "锁定积分",
                    "type": "integer"
                },
                "points": {
                    "description": "剩余积分",
                    "type": "integer"
                }
            }
        },
...
        "model.SettingQuestion": {
            "type": "object",
            "properties": {
                "answer": {
                    "description": "解答",
                    "type": "string"
                },
                "question": {
                    "description": "问题",
                    "type": "string"
                }
            }
        },
        "model.SettingRiderApp": {
            "type": "object",
            "properties": {
                "assistanceFee": {
                    "description": "救援费用",
                    "type": "number"
                },
                "reserveDuration": {
                    "description": "预约最长时间(分钟)",
                    "type": "integer"
                }
            }
        },
...
        "model.WalletOverview": {
            "type": "object",
            "properties": {
                "balance": {
                    "description": "账户余额",
                    "type": "number"
                },
                "coupons": {
                    "description": "可使用优惠券数量",
                    "type": "integer"
                },
                "deposit": {
                    "description": "已缴纳押金",
                    "type": "number"
                },
                "points": {
                    "description": "积分数量",
                    "type": "integer"
                }
            }
        }
    }
}